Code refactor
Class name changes (and filenames)
Manager -> SceneManager
TBD: Scene, all Scene elements to be inherited from the class Entity,
generic shader class

+++ XBeat/Input/InputManager.cpp
@@ -175,7 +175,7 @@ void Manager::ProcessInput()
 
 	if (!FAILED(result)) {
 		for (DWORD i = 0; i < items; i++) {
-			pressedKeys[keyChanges[i].dwOfs] = (keyChanges[i].dwData & 0x80) != 0 ? CallbackInfo::OnKeyDown : CallbackInfo::OnKeyUp;
+			pressedKeys[(uint8_t)keyChanges[i].dwOfs] = (keyChanges[i].dwData & 0x80) != 0 ? CallbackInfo::OnKeyDown : CallbackInfo::OnKeyUp;
 		}
 	}
